requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.1.1
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   channels
    #   daphne
    #   django
    #   django-countries
attrs==23.1.0
    # via
    #   -r requirements.in
    #   automat
    #   jsonschema
    #   service-identity
    #   twisted
autobahn==23.6.1
    # via
    #   -r requirements.in
    #   daphne
automat==22.10.0
    # via
    #   -r requirements.in
    #   twisted
babel==2.12.1
    # via -r requirements.in
beautifulsoup4==4.12.2
    # via
    #   -r requirements.in
    #   bs4
billiard==4.1.0
    # via
    #   -r requirements.in
    #   celery
bs4==0.0.1
    # via -r requirements.in
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
celery==5.3.0
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2023.5.7
    # via
    #   -r requirements.in
    #   requests
cffi==1.15.1
    # via
    #   -r requirements.in
    #   cryptography
channels==4.0.0
    # via -r requirements.in
charset-normalizer==3.1.0
    # via
    #   -r requirements.in
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via
    #   -r requirements.in
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements.in
    #   celery
click-repl==0.2.0
    # via
    #   -r requirements.in
    #   celery
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
    #   pytest
constantly==15.1.0
    # via
    #   -r requirements.in
    #   twisted
cryptography==41.0.1
    # via
    #   -r requirements.in
    #   autobahn
    #   pyopenssl
    #   service-identity
    #   social-auth-core
daphne==4.0.0
    # via -r requirements.in
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
django==4.2.1
    # via
    #   -r requirements.in
    #   channels
    #   django-appconf
    #   django-celery-results
    #   django-cors-headers
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   jsonfield
    #   social-auth-app-django
django-appconf==1.0.5
    # via
    #   -r requirements.in
    #   django-compressor
django-celery-results==2.5.1
    # via -r requirements.in
django-compressor==4.3.1
    # via -r requirements.in
django-cors-headers==4.1.0
    # via -r requirements.in
django-countries==7.5.1
    # via -r requirements.in
django-phonenumber-field==7.1.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.0
    # via -r requirements.in
execnet==1.9.0
    # via
    #   -r requirements.in
    #   pytest-xdist
fire==0.5.0
    # via
    #   -r requirements.in
    #   rav
hyperlink==21.0.0
    # via
    #   -r requirements.in
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   -r requirements.in
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via
    #   -r requirements.in
    #   twisted
inflection==0.5.1
    # via -r requirements.in
iniconfig==2.0.0
    # via
    #   -r requirements.in
    #   pytest
jsonfield==2.1.1
    # via -r requirements.in
jsonschema==4.17.3
    # via -r requirements.in
kombu==5.3.0
    # via
    #   -r requirements.in
    #   celery
markdown-it-py==2.2.0
    # via
    #   -r requirements.in
    #   rich
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
packaging==23.1
    # via
    #   -r requirements.in
    #   build
    #   pytest
phonenumbers==8.13.13
    # via -r requirements.in
pillow==9.5.0
    # via -r requirements.in
pip-tools==6.13.0
    # via -r requirements.in
pluggy==1.0.0
    # via
    #   -r requirements.in
    #   pytest
prompt-toolkit==3.0.38
    # via
    #   -r requirements.in
    #   click-repl
psycopg2==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via
    #   -r requirements.in
    #   service-identity
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pygments==2.15.1
    # via
    #   -r requirements.in
    #   rich
pyjwt==2.7.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pyopenssl==23.2.0
    # via
    #   -r requirements.in
    #   twisted
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pyrsistent==0.19.3
    # via
    #   -r requirements.in
    #   jsonschema
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-xdist
pytest-django==4.5.2
    # via -r requirements.in
pytest-xdist==3.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   celery
python-decouple==3.8
    # via -r requirements.in
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2023.3
    # via
    #   -r requirements.in
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.in
    #   rav
rav==0.0.8
    # via -r requirements.in
rcssmin==1.1.1
    # via
    #   -r requirements.in
    #   django-compressor
requests==2.31.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   social-auth-core
rich==13.4.1
    # via
    #   -r requirements.in
    #   rav
rjsmin==1.2.1
    # via
    #   -r requirements.in
    #   django-compressor
service-identity==21.1.0
    # via
    #   -r requirements.in
    #   twisted
six==1.16.0
    # via
    #   -r requirements.in
    #   automat
    #   click-repl
    #   fire
    #   jsonfield
    #   python-dateutil
    #   service-identity
social-auth-app-django==5.2.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.4.2
    # via
    #   -r requirements.in
    #   social-auth-app-django
soupsieve==2.4.1
    # via
    #   -r requirements.in
    #   beautifulsoup4
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
termcolor==2.3.0
    # via
    #   -r requirements.in
    #   fire
twisted[tls]==22.10.0
    # via
    #   -r requirements.in
    #   daphne
twisted-iocpsupport==1.0.3
    # via
    #   -r requirements.in
    #   twisted
txaio==23.1.1
    # via
    #   -r requirements.in
    #   autobahn
typing-extensions==4.6.3
    # via
    #   -r requirements.in
    #   django-countries
    #   twisted
tzdata==2023.3
    # via
    #   -r requirements.in
    #   celery
    #   django
uritemplate==4.1.1
    # via -r requirements.in
urllib3==2.0.3
    # via
    #   -r requirements.in
    #   requests
vine==5.0.0
    # via
    #   -r requirements.in
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via
    #   -r requirements.in
    #   prompt-toolkit
wheel==0.40.0
    # via pip-tools
zope-interface==6.0
    # via
    #   -r requirements.in
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
